player controller bounce major fixes
I think I have a better understanding on the issues we are facing with bounce issues. They are caused by the disorganized hierarchy structure of the prefabs. Colliders added as empties, tagged as Barriers and having the Barrier Layer assigned, are accurate, however are not children of the proper object. I found that it works best for detection when the collider is child to the fbx model asset, rather than a sibling to it in the directory. I also added child collider boxes tagged with barrier and on the barrier layer to numerous bars in the prefabs.