continued work on the player look to grab/ raycast
I am trying to determine a better hierarchy for the tags in the raycasting to ensure that bars are the highest. created two different raycasts to store the hits, if we get a bar hit in our shotgun blast, it stores it. If we get a wall, door, lever, it stores them as fall backs. Therefore, this allows us to process them later. If there is a bar in the ray cast, sick, we are showing the grabber. If we don't, then the other stuff can be fall backs. it's working better, but still has issues with detecting a bar closest to the player. as well I haven't gotten implementation of the update closest bar in view.