Caught myself going down a rabbit hole
Refined the grab and push off wall logic, as well all the lockdown event stuff all works. However, the pickup script is un-optomized and causes frame dips on my screen whenever I grab a block. I tried to see if I could get the grabber to show up but this seems like its own, big issue for a card.

 I believe in the future the pick up script should be implemented directly into the uimanager. Therfore we can create arrays that serve as buffers.

rn the pick up works, but there is no grabber as everything I did to add stuff to fix was useless and caused issues other places. better for a unique card to address,